+++ src/library/Hasql/Errors.hs
@@ -235,6 +235,8 @@ instance IsError SessionError where
       toMessage execErr
     ConnectionSessionError {} ->
       "Connection error"
+    ClientRejectionSessionError {} ->
+      "Client rejected the request"
     DriverSessionError {} ->
       "Driver error"
     MissingTypesSessionError {} ->
@@ -253,6 +255,8 @@ instance IsError SessionError where
       ("sql", sql) : toDetails execErr
     ConnectionSessionError reason ->
       [("reason", reason)]
+    ClientRejectionSessionError reason ->
+      [("reason", reason)]
     DriverSessionError reason ->
       [("reason", reason)]
     MissingTypesSessionError missingTypes ->
@@ -265,6 +269,7 @@ instance IsError SessionError where
 
   isTransient = \case
     ConnectionSessionError _ -> True
+    ClientRejectionSessionError _ -> False
     StatementSessionError _ _ _ _ _ statementError -> isTransient statementError
     ScriptSessionError _ serverError -> isTransient serverError
     DriverSessionError {} -> False
@@ -274,5 +279,6 @@ instance IsError SessionError where
     StatementSessionError _ _ _ _ _ statementError -> toSqlState statementError
     ScriptSessionError _ serverError -> toSqlState serverError
     ConnectionSessionError {} -> Nothing
+    ClientRejectionSessionError {} -> Nothing
     DriverSessionError {} -> Nothing
     MissingTypesSessionError {} -> Nothing